removed unnecessary /MainBody and renamed /Cards = /MainBody
