+++ hw3/03-react/src/components/houses/DonutChart.jsx
@@ -18,9 +18,9 @@ const backgroundColors = [
   "rgba(49, 70, 89, 0.9)",
   getRGBA(0.8),
   "#181a1f",
-  getRGBA(0.8),
+  getRGBA(0.9),
   "#923734",
-  getRGBA(0.8),
+  getRGBA(0.9),
   "#ffdf8c",
   getRGBA(0.8),
   "#525949",
