+++ hw3/03-react/src/components/houses/styles/house.css
@@ -6,3 +6,9 @@ canvas {
   max-height: 90%;
   max-width: auto;
 }
+
+.card-container {
+  overflow-x: hidden;
+  overflow-y: auto;
+  text-align: justify;
+}
